add anti-aliasing to preprocessing and convert convae to half-precision

+++ src/conv_AE_2D.py
@@ -16,7 +16,7 @@ HEIGHT = 227
 class ConvAE2d(object):
     def __init__(self, tvol, alpha, batch_size, lambd):
         self.tvol = tvol
-        self.x_ = tf.placeholder(tf.float32, [None, self.tvol, HEIGHT, WIDTH, NCHANNELS])
+        self.x_ = tf.placeholder(tf.float16, [None, self.tvol, HEIGHT, WIDTH, NCHANNELS])
         self.phase = tf.placeholder(tf.bool, name='is_training')
 
         self.batch_size = batch_size
